Eye - Python Code
Python code that uses machine learning (neural networks) and computer
vision to classify images based on identified features. The
classification is stored within the image file (as tags). This allows to
search images based on features (e.g. searching for ‘cat’ will return
images with cat). The program is written in Python and uses Keras with
TensorFlow.